+++ src/index.js
@@ -3,3 +3,4 @@ export { inspectProject, writeArtifacts } from "./manifest.js";
 export { buildProject } from "./build.js";
 export { listSkills } from "./skills.js";
 export { docsPath, listDocs } from "./docs.js";
+export { listTemplates, getTemplateScaffold } from "./templates.js";